server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
gitpython==3.1.43
python-dotenv==1.0.1
PyGithub==2.4.0
